# perf: Hashing account stage, full state