--- a/crates/stages/src/stages/hashing_account.rs
+++ b/crates/stages/src/stages/hashing_account.rs
@@ -181,8 +181,6 @@ impl<DB: Database> Stage<DB> for AccountHashingStage {
         // account otherwise take changesets aggregate the sets and apply hashing to
         // AccountHashing table. Also, if we start from genesis, we need to hash from scratch, as
         // genesis accounts are not in changeset.
-        let time = std::time::Instant::now();
-        println!("\n\n");
         if to_transition - from_transition > self.clean_threshold || stage_progress == 0 {
             let mut checkpoint = self.get_checkpoint(tx)?;
 
@@ -203,9 +201,8 @@ impl<DB: Database> Stage<DB> for AccountHashingStage {
             let next_address = {
                 let mut accounts_cursor = tx.cursor_read::<tables::PlainAccountState>()?;
 
-                // channels used to return result of sender recovery.
+                // channels used to return result of account hashing
                 let mut channels = Vec::new();
-                println!("START READING ACCOUNTS: {}ms", time.elapsed().as_millis());
                 for chunk in &accounts_cursor
                     .walk(start_address)?
                     .take(self.commit_threshold as usize)
@@ -216,32 +213,24 @@ impl<DB: Database> Stage<DB> for AccountHashingStage {
                     channels.push(rx);
 
                     let chunk = chunk.collect::<Result<Vec<_>, _>>()?;
-                    println!("READ CHUNK: {}ms", time.elapsed().as_millis());
-                    // Spawn the sender recovery task onto the global rayon pool
-                    // This task will send the results through the channel after it recovered the
-                    // senders.
+                    // Spawn the hashing task onto the global rayon pool
                     rayon::spawn(move || {
                         for (address, account) in chunk.into_iter() {
                             let _ = tx.send((keccak256(address), account));
                         }
                     });
                 }
-                println!("READ ALL ACCOUNTS: {}ms", time.elapsed().as_millis());
                 let mut hashed_batch = Vec::with_capacity(self.commit_threshold as usize);
 
-                // Iterate over channels and append the hashes account in the order that they are
-                // received.
+                // Iterate over channels and append the hashed accounts.
                 for mut channel in channels {
                     while let Some(hashed) = channel.recv().await {
                         hashed_batch.push(hashed);
                     }
                 }
-                println!("RECEIVED ALL ACCOUNTS: {}ms", time.elapsed().as_millis());
 
-                // sort it all
+                // sort it all in parallel
                 hashed_batch.par_sort_unstable_by(|a, b| a.0.cmp(&b.0));
-
-                println!("SORTED ALL ACCOUNTS: {}ms", time.elapsed().as_millis());
 
                 let mut hashed_account_cursor = tx.cursor_write::<tables::HashedAccount>()?;
 
@@ -255,8 +244,6 @@ impl<DB: Database> Stage<DB> for AccountHashingStage {
                         .into_iter()
                         .try_for_each(|(k, v)| hashed_account_cursor.insert(k, v))?;
                 }
-
-                println!("INSERTED ALL ACCOUNTS: {}ms", time.elapsed().as_millis());
 
                 // next key of iterator
                 accounts_cursor.next()?